updating

+++ stm32f4-HAL/Inc/stm32f4xx_hal_conf.h
@@ -92,7 +92,7 @@
   *        (when HSE is used as system clock source, directly or through the PLL).  
   */
 #if !defined  (HSE_VALUE) 
-  #define HSE_VALUE              25000000U /*!< Value of the External oscillator in Hz */
+  #define HSE_VALUE            ((uint32_t) 25000000u) /*!< Value of the External oscillator in Hz */
 #endif /* HSE_VALUE */
 
 #if !defined  (HSE_STARTUP_TIMEOUT)
@@ -105,7 +105,7 @@
   *        (when HSI is used as system clock source, directly or through the PLL). 
   */
 #if !defined  (HSI_VALUE)
-  #define HSI_VALUE              16000000U /*!< Value of the Internal oscillator in Hz */
+  #define HSI_VALUE        ((uint32_t) 16000000u) /*!< Value of the Internal oscillator in Hz */
 #endif /* HSI_VALUE */
 
 /**
